analytics/spark/silver-curation/requirements.txt
pyspark==3.5.1
pyarrow>=14.0.1
boto3>=1.34.0
fastavro>=1.9.5
pyyaml>=6.0.1
